Abort script much earlier if BI_API_KEY is unset
The redesigned exampapers page takes a lot longer to load the search results.
Until this commit, not setting BI_API_KEY would only error after trying to
access BI with the first PDF file, which could be up to a minute after starting
the script.

This commit ensures that the environment variable is checked on startup for
immediate feedback if you forgot.